+++ scm-ui/ui-components/src/index.ts
@@ -96,6 +96,7 @@ export * from "./navigation";
 export * from "./repos";
 export * from "./table";
 export * from "./toast";
+export * from "./popover";
 
 export {
   File,